+++ CHANGELOG.md
@@ -18,6 +18,7 @@ In order to read more about upgrading and BC breaks have a look at the [UPGRADE 
 
 ### Fixed
 
++ [#203](https://github.com/luyadev/luya-module-cms/issues/203) Show error page when NotFound exception is thrown.
 + [#201](https://github.com/luyadev/luya-module-cms/issues/201) Fix issued where language switcher widget does not prefix host info (defined by hostInfoMapping). 
 + [#193](https://github.com/luyadev/luya-module-cms/pull/193) Prefix certain application env folders when running block importer in order to make sure windows system paths work.
 + [#194](https://github.com/luyadev/luya-module-cms/issues/194) Updated CSS to unify spaces between page languages and treenav